// client/src/app/Models/Institute.ts
export interface InstituteInterface{
    institute_code?: number;
    institute_name?: string;
    ubication?: string;
    type?: string;
    image?: string;
    pathImage?: string;
}
export interface InstitutePrepareInterface{
    institute_code?: number;
    institute_name?: string;
    ubication?: string;
    type?: string;
    image?: string;
}